requirements.txt
openai==1.51.0
httpx==0.27.2
python-docx==1.1.2
tkinter-tooltip==2.1.0
requests==2.31.0
python-dotenv==1.0.0
Flask==3.0.3

